+++ server/json_rpc.go
@@ -29,10 +29,9 @@ import (
 	"github.com/cosmos/cosmos-sdk/server/types"
 	ethlog "github.com/ethereum/go-ethereum/log"
 	ethrpc "github.com/ethereum/go-ethereum/rpc"
-	"github.com/evmos/ethermint/app"
+	"github.com/evmos/ethermint/app/ante"
 	"github.com/evmos/ethermint/rpc"
 	"github.com/evmos/ethermint/rpc/stream"
-
 	"github.com/evmos/ethermint/server/config"
 	ethermint "github.com/evmos/ethermint/types"
 )
@@ -40,7 +39,7 @@ import (
 const MaxRetry = 6
 
 type AppWithPendingTxStream interface {
-	RegisterPendingTxListener(listener app.PendingTxListener)
+	RegisterPendingTxListener(listener ante.PendingTxListener)
 }
 
 // StartJSONRPC starts the JSON-RPC server